Backend/old_code/requirements_old.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-RESTX==1.3.0
Flask-CORS==4.0.0
Flask-Mail==0.9.1
python-dotenv==1.0.0
Werkzeug==3.0.1
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
msal==1.28.0
PyJWT==2.8.0
cryptography==41.0.7
requests==2.31.0

# Google Cloud Vision OCR
google-cloud-vision==3.4.5

# Celery & Redis for async processing
celery==5.3.4
redis==5.0.1
kombu==5.3.4
billiard==4.2.0
vine==5.1.0

# Image processing
Pillow==10.1.0
pdf2image==1.16.3
PyPDF2==3.0.1

